Add files via upload
The homework directions were not completely clear that we needed the data aggregation to appear on each worksheet so I created a separate worksheet to aggregate ALL of the data.  The multiple recurrences of the stock tickers as shown in the second image represent individual years (filtered on the displayed).